requirements.txt
pandas==2.3.3
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
requests==2.31.0
python-dotenv
requests-cache
flask
beautifulsoup4